requirements.txt
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
docutils==0.17.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.6.0
incremental==21.3.0
inflection==0.5.1
itemadapter==0.3.0
itemloaders==1.0.4
jmespath==0.10.0
keyring==23.0.1
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
matplot==0.1.9
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.0
packaging==21.0
pandas==1.2.5
parsel==1.6.0
pdfminer.six==20201018
Pillow==8.3.0
pkginfo==1.7.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.9.0
pyloco==0.0.139
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyPdf==1.13
python-dateutil==2.8.1
pytz==2021.1
pywin32-ctypes==0.2.0
queuelib==1.6.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
Scrapy==2.5.0
selenium==3.141.0
service-identity==21.1.0
SimpleWebSocketServer==0.1.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
tqdm==4.61.1
twine==3.4.1
Twisted==21.7.0
twisted-iocpsupport==1.0.1
typing==3.7.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
ushlex==0.99.1
w3lib==1.22.0
webencodings==0.5.1
websocket-client==1.1.0
zipp==3.5.0
zope.interface==5.4.0
